move storageapp to fs
